+++ build.sh
@@ -19,7 +19,7 @@ echo "Building for $ARCH"
 
 function check_deps() {
   echo "Checking dependencies"
-  for DEP in wget tar gzip cpio dd losetup sfdisk mkfs.vfat mkfs.ext4 sudo install sed patch
+  for DEP in wget tar gzip dd losetup sfdisk mkfs.vfat mkfs.ext4 sudo install
   do
     if ! command -v "$DEP" > /dev/null
     then
@@ -48,30 +48,43 @@ install_tmp() {
   cd "$TMPDIR" || exit 1
 }
 
-build_stage1() {
-  echo "myMPDos build stage 1: Download"
-  install_tmp
-  if [ ! -f "$TMPDIR/$NETBOOT_ARCHIVE" ]
-  then
-    echo "Getting $NETBOOT_ARCHIVE"
-    wget -q "${ALPINE_MIRROR}/v${ALPINE_MAJOR_VERSION}/releases/${ARCH}/$NETBOOT_ARCHIVE" \
-      -O "$TMPDIR/$NETBOOT_ARCHIVE"
-  fi
-  if [ ! -d "$TMPDIR/netboot" ]
-  then
-    install -d "$TMPDIR/netboot"
-    if ! tar -xzf "$TMPDIR/$NETBOOT_ARCHIVE" -C "$TMPDIR/netboot"
-    then
-      echo "Can not extract $NETBOOT_ARCHIVE"
-      rm -f "$NETBOOT_ARCHIVE"
-      exit 1
-    fi
-  fi
+register_binfmt() {
+  docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
+}
 
+build_packages() {
+  # This will be the build volume
+  mkdir -p "$TMPDIR/build"
+  mkdir -p "$TMPDIR/apk-cache"
+  # Archive signing key
+  if [ -f "$STARTPATH/keys/abuild.tgz" ]
+  then
+    echo "Using keys for public repository"
+    cp "$STARTPATH/keys/abuild.tgz" "$TMPDIR/build/"
+  elif [ -f "$STARTPATH/apks/abuild.tgz" ]
+  then
+    echo "Using private build keys"
+    cp "$STARTPATH/apks/abuild.tgz" "$TMPDIR/build/"
+  else
+    echo "No saved abuild.tgz found"
+  fi
+  docker run --rm \
+    -v "$TMPDIR/apk-cache":"/var/cache/apk":Z \
+    -v "${STARTPATH}/mympdos/build":"/mympdos":Z \
+    -v "$TMPDIR/build":"/build":Z \
+    -v "${STARTPATH}/repository/$ARCH":"/build/packages/package/$ARCH":Z \
+    --platform linux/arm64/v8 -it arm64v8/alpine \
+    /bin/sh
+    #/mympdos/build.sh
+}
+
+build_image() {
+  echo "Create myMPDos image"
+  install_tmp
   if [ ! -f "$TMPDIR/$ARCHIVE" ]
   then
     echo "Getting $ARCHIVE"
-    wget -q "${ALPINE_MIRROR}/v${ALPINE_MAJOR_VERSION}/releases/${ARCH}/$ARCHIVE" \
+    wget "${ALPINE_MIRROR}/v${ALPINE_MAJOR_VERSION}/releases/${ARCH}/$ARCHIVE" \
       -O "$TMPDIR/$ARCHIVE"
     if ! tar -tzf "$TMPDIR/$ARCHIVE" > /dev/null
     then
@@ -80,113 +93,6 @@ build_stage1() {
       exit 1
     fi
   fi
-}
-
-build_stage2() {
-  echo "myMPDos build stage 2: Create build image"
-  install_tmp
-  dd if=/dev/zero of="$TMPDIR/$BUILDIMAGE" bs=1M count="$IMAGESIZEBUILD"
-  sfdisk "$TMPDIR/$BUILDIMAGE" <<< "1, ${BOOTPARTSIZEBUILD}, b, *"
-  sfdisk -a "$TMPDIR/$BUILDIMAGE" <<< ","
-
-  LOOP=$(sudo losetup --partscan --show -f "$BUILDIMAGE")
-  [ "$LOOP" = "" ] && exit 1
-  sudo mkfs.vfat "${LOOP}p1"
-  sudo mkfs.ext4 "${LOOP}p2"
-  install -d "$TMPDIR/mnt"
-  sudo mount -ouid="$BUILDUSER" "${LOOP}p1" "$TMPDIR/mnt" || exit 1
-  if ! tar -xzf "$TMPDIR/$ARCHIVE" -C "$TMPDIR/mnt"
-  then
-    echo "Extracting $TMPDIR/$ARCHIVE failed"
-    exit 1
-  fi
-  cp "$TMPDIR/netboot/boot/$MODLOOP" "$TMPDIR/mnt/boot"
-
-  echo "Copy build scripts"
-  install -d "$TMPDIR/mnt/mympdos"
-  find "$TMPDIR/mnt/mympdos/" -name \*~ -delete
-  cp -r "$STARTPATH"/mympdos/build/* "$TMPDIR/mnt/mympdos"
-
-  echo "Copy existing packages"
-  install -d "$TMPDIR/mnt/mympdos-apks"
-  if [ -f "$STARTPATH/repository/$ARCH/APKINDEX.tar.gz" ]
-  then
-    cp "$STARTPATH/repository/$ARCH/"*.apk "$TMPDIR/mnt/mympdos-apks/"
-    cp "$STARTPATH/repository/$ARCH/APKINDEX.tar.gz" "$TMPDIR/mnt/mympdos-apks/"
-  else
-    echo "No existing packages found"
-  fi
-  if [ -f "$STARTPATH/keys/abuild.tgz" ]
-  then
-    echo "Using keys for public repository"
-    cp "$STARTPATH/keys/abuild.tgz" "$TMPDIR/mnt/mympdos/"
-  elif [ -f "$STARTPATH/apks/abuild.tgz" ]
-  then
-    echo "Using private build keys"
-    cp "$STARTPATH/apks/abuild.tgz" "$TMPDIR/mnt/mympdos/"
-  else
-    echo "No saved abuild.tgz found"
-  fi
-  date +%s > "$TMPDIR/mnt/date"
-  umount_retry "$TMPDIR/mnt" || exit 1
-  sudo losetup -d "${LOOP}"
-
-  echo "Patching initramfs"
-  cd "$TMPDIR/netboot" || exit 1
-  #rm -f "init"
-  #gzip -dc "boot/$INITRAMFS" | cpio -id init
-  if ! cp "$STARTPATH/mympdos/netboot/init.mympd" init
-  then
-    echo "Patching netboot init failed"
-    exit 1
-  fi
-  echo ./init | cpio -H newc -o | gzip >> "$TMPDIR/netboot/boot/$INITRAMFS"
-}
-
-build_stage3() {
-  echo "myMPDos build stage 3: Starting build"
-  install_tmp
-  $QEMU \
-    -M virt -m "$BUILDRAM" -cpu "$CPU" -smp "$BUILDCPUS" \
-    -kernel "$TMPDIR/netboot/boot/$KERNEL" -initrd "$TMPDIR/netboot/boot/$INITRAMFS" \
-    -append "console=ttyAMA0 ip=dhcp" \
-    -nographic \
-    -drive "file=$TMPDIR/${BUILDIMAGE},format=raw" \
-    -nic user,id=mynet0,net=192.168.76.0/24,dhcpstart=192.168.76.9
-}
-
-build_stage4() {
-  echo "myMPDos build stage 4: Saving packages"
-  install_tmp
-  if [ -d "$STARTPATH/apks" ]
-  then
-    BACKUPDATE=$(stat -c"%Y" "$STARTPATH/apks")
-    BACKUPDIR="$STARTPATH"/apks.$(date -d@"$BACKUPDATE" +%Y%m%d_%H%M)
-    mv "$STARTPATH/apks" "$BACKUPDIR"
-  fi
-  install -d "$STARTPATH/apks/$ARCH"
-  LOOP=$(sudo losetup --partscan --show -f "$TMPDIR/$BUILDIMAGE")
-  sudo mount -text4 "${LOOP}p2" "$TMPDIR/mnt" || exit 1
-  if [ -f "$TMPDIR/mnt/build/abuild.tgz" ]
-  then
-    cp "$TMPDIR/mnt/build/abuild.tgz" "$STARTPATH/apks/"
-  else
-    echo "No abuild.tgz found"
-  fi
-  if [ -f "$TMPDIR/mnt/build/packages/package/${ARCH}/APKINDEX.tar.gz" ]
-  then
-    cp "$TMPDIR"/mnt/build/packages/package/"${ARCH}"/* "$STARTPATH/apks/$ARCH/"
-    cp "$TMPDIR"/mnt/build/packages/package/"${ARCH}"/* "$STARTPATH/repository/$ARCH/"
-  else
-    echo "No APKINDEX.tar.gz found"
-  fi
-  umount_retry "$TMPDIR/mnt" || exit 1
-  sudo losetup -d "${LOOP}"
-}
-
-build_stage5() {
-  echo "myMPDos build stage 5: Create image"
-  install_tmp
   dd if=/dev/zero of="$TMPDIR/$IMAGE" bs=1M count="$IMAGESIZE"
   sfdisk "$TMPDIR/$IMAGE" <<< "1, ${BOOTPARTSIZE}, b, *"
 
@@ -290,39 +196,21 @@ umountbuild() {
 }
 
 case "$2" in
-  private|p)
+  private)
     PRIVATEIMAGE="true";;
   *)
     PRIVATEIMAGE="false";;
 esac
 
 case "$1" in
-  stage1|1)
-    check_deps
-    build_stage1
+  register|r)
+    register_binfmt
     ;;
-  stage2|2)
-    check_deps
-    build_stage2
+  packages|p)
+    build_packages
     ;;
-  stage3|3)
+  image|i)
     check_deps
-    build_stage3
-    ;;
-  stage4|4)
-    check_deps
-    build_stage4
-    ;;
-  stage5|5)
-    check_deps
-    build_stage5
-    ;;
-  build|b)
-    check_deps
-    build_stage1
-    build_stage2
-    build_stage3
-    build_stage4
     build_stage5
     ;;
   umountbuild|u)
@@ -332,21 +220,17 @@ case "$1" in
     cleanup
     ;;
   *)
-    echo "Usage: $0 (b|1|2|3|4|5|c|u) [private|public]"
+    echo "Usage: $0 (r|p|i|c|u) [private|public]"
     echo ""
-    echo "  build|b:        runs all stages"
-    echo "  stage1|1:       downloads and extracts all needed sources"
-    echo "  stage2|2:       creates the build image"
-    echo "  stage3|3:       starts the build image"
-    echo "  stage4|4:       copies the packages from build into apks"
-    echo "  stage5|5:       creates the image"
+    echo "  register|r:     Registers binfmt with QEMU for different architectures"
+    echo "  packages|p:     Builds the packages"
+    echo "  image|i:        Creates the image"
     echo ""
     echo "  cleanup|c:      cleanup things"
     echo "  umountbuild|u:  removes dangling mounts and loop devices"
     echo ""
-    echo "  private|p:      creates a image with a productive bootstrap.txt file"
+    echo "  private:        creates a image with a productive bootstrap.txt file"
     echo "  public:         creates a image with sample bootstrap.txt files (default)"
+    exit 2
     ;;
 esac
-
-exit 0