+++ updates/1.8.2/update/update.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+export V_MAJOR=3
+export V_MINOR=24
+export V_POINT=1
+export CHECKSUM=f50672e626f234e13686eb7534292e5b7c0a639986d8afd87c25526bc7a75882
+alpine-upgrade.sh