+++ CHANGELOG.md
@@ -6,8 +6,11 @@ https://github.com/jcorporation/myMPDos/
 
 ## myMPDos v1.9.0
 
+This release is build up on Alpine Linux 3.24.1.
+
 ### Changelog
 
+- Alpine Linux 3.24.1
 - Move mympd doas config from bootstrap script to mympdos-base package
 
 ***